TestUpdateTimestamp
